Set up custom two-way binding
In this commit, was made and seen a way of use a custom two-way binding
with input and output decorators. The two decorators created started
with the same name which is "size", and the size property in the
app.component.html was putted inside [()] notation to allow the feature
work.